# Improve Polynomial operation
## Benchmark results
```jl
using BenchmarkTools
using QRCoders
using QRCode.Polynomial: Poly as OriPoly

randpoly(n::Int) = Poly([rand(0:255, n-1)..., rand(1:255)])
f1 = randpoly(100);
f2 = OriPoly(f1.coeff);
@btime f1 * f1;
@btime f2 * f2;

# 9.219 μs (2 allocations: 1.78 KiB)
# 325.375 μs (501 allocations: 555.11 KiB)
```

For the encoding part, polynomial operation is used only in `geterrorcorrection`.
```jl
exportqrcode =1> qrcode =1> encodemessage
=at most 59> getecblock =1> geterrorcorrection
```

So it decreases little of the efficiency comparing to the other operations.

Current result
| msg | cost |
| --- | --- |
| "Hello world!" | 129.871 μs (270 allocations: 19.34 KiB) |
| "0123456789" ^ 30 | 1.223 ms (1068 allocations: 96.20 KiB) |
| "HELLO WORLD" ^ 30 | 2.036 ms (1655 allocations: 157.61 KiB) |
| "HELLO WORLD" ^ 190 | 11.962 ms (8930 allocations: 882.53 KiB) |

Previous one
| input string | cost |
| --- | --- |
|  "Hello world!" | `142.909 μs (370 allocations: 32.23 KiB)`|
|  "0123456789" ^ 30 | `1.454 ms (1982 allocations: 291.95 KiB)` |
| "HELLO WORLD" ^ 30 | `2.495 ms (3200 allocations: 562.52 KiB)` |
|  "HELLO WORLD" ^ 190 | `14.310 ms (17825 allocations: 3.01 MiB)` |


However, **Efficiency of polynomial operations are important in QRDecoders.jl, since they are used frequently.**

---

Original implement:
```jl
function *(a::Poly, b::Poly)::Poly
    return sum([ c * (a << (p - 1)) for (p, c) in enumerate(b.coeff)])
end
```
Here `c * Poly`, `a << Int` and `Poly + Poly` will create new polynomials.

New one:
```jl
function *(a::Poly, b::Poly)::Poly
    prodpoly = Poly(zeros(Int, length(a) + length(b) - 1))
    @inbounds for (i, c1) in enumerate(a.coeff), (j, c2) in enumerate(b.coeff)
        prodpoly.coeff[i + j - 1] ⊻= mult(c2, c1) # column first
    end
    return prodpoly
end
```